Reference

RestrictedBoltzmannMachines.RBM — Method

RBM(white_rbm::WhiteRBM)

Returns an (unwhitened) RBM which neglects the affine transforms of white_rbm. The resulting model is not equivalent to the original white_rbm. To construct an equivalent model, use the function blacken(white_rbm) instead (see blacken).

WhitenedRBMs.WhiteRBM — Method

WhiteRBM(rbm)

Creates a WhiteRBM with identity transforms.

Base.one — Method

one(t::Affine)

Returns the identity affine transformation, which maps all points to themselves.

Base.zero — Method

zero(t::Affine)

Returns the zero affine transformation, which maps all points to zero.

WhitenedRBMs.BinaryWhiteRBM — Method

BinaryWhiteRBM(a, b, w, affine_v, affine_h)
BinaryWhiteRBM(a, b, w)

Construct a whitened RBM with binary visible and hidden units.

WhitenedRBMs.blacken — Method

blacken(white_rbm::WhiteRBM)

Constructs a plain RBM equivalent to the given white_rbm. The energies assigned by the two models differ by a constant amount,

\[E(v,h) - \tilde{E}(v,h) = \mathbf{a}^\top \mathbb{A}^\top \tilde{\mathbb{W}} \mathbb{B}\mathbf{b}\]

where $\tilde{E}(v,h)$ is the energy assigned by white_rbm and $E(v,h)$ is the energy assigned by the RBM constructed by this method.

This is the inverse operation of whiten, always returning an ordinary RBM.

To construct an RBM that simply neglects the transformations, call RBM(white_rbm) instead.

WhitenedRBMs.energy_shift — Method

energy_shift(rbm, affine_v, affine_h)

Computes the constant energy shift if the affine transformations were updated as given.

WhitenedRBMs.hidden_affine_from_inputs — Function

hidden_affine_from_inputs(white_rbm, data; wts = nothing, ϵ = 0)

Returns the affine transformation that standardizes the hidden inputs. Note that only the variances are scaled to unity. The correlation matrix is not diagonalized.

WhitenedRBMs.safe_whiten — Method

safe_whiten(white_rbm, affine_v, affine_h)

Like whiten(white_rbm, affine_v, affine_h), but ensures that the returned WhiteRBM is of the same type as white_rbm (in particular, preserving the types of the affine transformations).

WhitenedRBMs.shift_fields — Function

shift_fields(layer, offset)

Adds offset to the layer fields.

WhitenedRBMs.shift_fields! — Function

shift_fields!(layer, offset)

In-place version of shift_fields(layer, offset).

WhitenedRBMs.visible_affine_from_data — Function

visible_affine_from_data(white_rbm, data; wts = nothing, ϵ = 0)

Returns the affine transformation that whitens the visible data.

WhitenedRBMs.whiten — Method

whiten(rbm::RBM, affine_v, affine_h)

Constructs a WhiteRBM equivalent to the given rbm. The energies assigned by the two models differ by a constant amount,

\[E(v,h) - \tilde{E}(v,h) = \mathbf{a}^\top \mathbb{A}^\top \tilde{\mathbb{W}} \mathbb{B}\mathbf{b}\]

where $E(v,h)$ is the energy assigned by the original rbm, and $\tilde{E}(v,h)$ is the energy assigned by the returned WhiteRBM.

This is the inverse operation of blacken, always returning a WhiteRBM.

To construct a WhiteRBM that simply includes these transforms, call WhiteRBM(rbm, affine_v, affine_h) instead.

WhitenedRBMs.whitening_transform — Method

whitening_transform(μ, C)

Returns the Affine transform that whitens data with mean μ and covariance C.